r31 | lanceball | 2007-07-09 15:10:27 -0700 (Mon, 09 Jul 2007) | 1 line
Added a class to help with profiling the daemon.  The class will initialize a config, master and 2 slave databases.  Then run the PgJDBCBench class on the master.  Using the JProfiler plugin for IntelliJ IDEA makes it very easy to run this class and do all kinds of testing on the daemon while the profiler is running.


git-svn-id: file:///home/mgrosso/data/svn-elvis/codehause-import/branches/trunk-history@57 2e195dce-9def-4caf-bb7c-8c699b2d911e